requirements.txt
dash==2.17.1
pandas==2.2.2
pypdf==4.2.0
pdfplumber==0.11.4
python-docx==1.1.2
openpyxl==3.1.5
pillow==10.4.0
requests==2.32.3
beautifulsoup4==4.12.3
scikit-learn==1.5.1
python-dotenv==1.0.1
litellm
openai==1.40.0
matplotlib==3.9.0
